🚚(frontend) restructure SyllabusCourseRunsList widget
Within `SyllabusCourseRunsList`, several components were created at the folder
root that was not follow our architecture of other component. So we restrucutre
this widget in order to stick to file order convention.
