+++ src/frontend/js/widgets/SyllabusCourseRunsList/components/SyllabusSimpleCourseRunsList/index.tsx
