+++ src/frontend/js/widgets/SyllabusCourseRunsList/index.tsx
@@ -4,8 +4,8 @@ import { createPortal } from 'react-dom';
 import { CourseRun, Priority } from 'types';
 import { Course } from 'types/Course';
 import { computeStates } from 'utils/CourseRuns';
-import { SyllabusAsideList } from 'widgets/SyllabusCourseRunsList/SyllabusAsideList';
-import { SyllabusCourseRun } from 'widgets/SyllabusCourseRunsList/SyllabusCourseRun';
+import { SyllabusAsideList } from 'widgets/SyllabusCourseRunsList/components/SyllabusAsideList';
+import { SyllabusCourseRun } from 'widgets/SyllabusCourseRunsList/components/SyllabusCourseRun';
 import { DjangoCMSPluginsInit } from 'components/DjangoCMSTemplate';
 import { isJoanieEnabled } from 'api/joanie';
 import context from 'utils/context';
